+++ src/main/java/org/programmers/Application.java
@@ -2,17 +2,20 @@ package org.programmers;
 
 import org.programmers.Io.Console;
 import org.programmers.calculator.Calculator;
-import org.programmers.calculator.ExpressionEvaluator;
+import org.programmers.calculator.Calculate;
 import org.programmers.repository.CalRepository;
 import org.programmers.repository.CalRepositoryImpl;
+import org.programmers.validator.Validator;
 
 public class Application {
     public static void main(String[] args) {
         Console console = new Console();
-        ExpressionEvaluator expressionEvaluator = new ExpressionEvaluator();
+        Calculate calculate = new Calculate();
         CalRepository calRepository = new CalRepositoryImpl();
+        Validator validator = new Validator();
+        Calculator calculator = new Calculator(console, calculate, calRepository, validator);
 
-        Calculator calculator = new Calculator(console, expressionEvaluator, calRepository);
+
         calculator.run();
     }
 }
